+++ gork/zmemory.go
@@ -7,7 +7,7 @@ import (
 type ZMemory []byte
 type ZMemorySequential struct {
 	mem *ZMemory
-	pos uint16
+	pos uint32
 }
 
 func NewZMemory(mem []byte) *ZMemory {
@@ -16,17 +16,17 @@ func NewZMemory(mem []byte) *ZMemory {
 	return zmem
 }
 
-func (zmem *ZMemory) ByteAt(addr uint16) byte {
+func (zmem *ZMemory) ByteAt(addr uint32) byte {
 	return (*zmem)[addr]
 }
 
-func (zmem *ZMemory) WordAt(addr uint16) uint16 {
+func (zmem *ZMemory) WordAt(addr uint32) uint16 {
 	// Big Endian
 	return (uint16((*zmem)[addr]) << 8) |
 		(uint16((*zmem)[addr+1]))
 }
 
-func (zmem *ZMemory) UInt32At(addr uint16) uint32 {
+func (zmem *ZMemory) UInt32At(addr uint32) uint32 {
 	// Big Endian
 	return (uint32((*zmem)[addr]) << 24) |
 		(uint32((*zmem)[addr+1]) << 16) |
@@ -34,16 +34,16 @@ func (zmem *ZMemory) UInt32At(addr uint16) uint32 {
 		uint32((*zmem)[addr+3])
 }
 
-func (zmem *ZMemory) WriteByteAt(addr uint16, val byte) {
+func (zmem *ZMemory) WriteByteAt(addr uint32, val byte) {
 	(*zmem)[addr] = val
 }
 
-func (zmem *ZMemory) WriteWordAt(addr uint16, val uint16) {
+func (zmem *ZMemory) WriteWordAt(addr uint32, val uint16) {
 	(*zmem)[addr] = byte(val >> 8)
 	(*zmem)[addr+1] = byte(val & 0X00FF)
 }
 
-func (zmem *ZMemory) GetSequential(addr uint16) *ZMemorySequential {
+func (zmem *ZMemory) GetSequential(addr uint32) *ZMemorySequential {
 	return &ZMemorySequential{zmem, addr}
 }
 
@@ -77,7 +77,7 @@ func (zmem *ZMemorySequential) ReadUint32() uint32 {
 	return tmp
 }
 
-func (zmem *ZMemory) DecodeZStringAt(addr uint16, header *ZHeader) string {
+func (zmem *ZMemory) DecodeZStringAt(addr uint32, header *ZHeader) string {
 	return zmem.GetSequential(addr).DecodeZString(header)
 }
 
@@ -110,7 +110,7 @@ func (zmem *ZMemorySequential) DecodeZString(header *ZHeader) string {
 				synonimFlag = false
 				synonim = (synonim - 1) * 64
 
-				tmpAddr := zmem.mem.WordAt(header.abbrTblPos+synonim+code*2) * 2
+				tmpAddr := uint32(zmem.mem.WordAt(uint32(header.abbrTblPos+synonim+code*2))) * 2
 				ret += zmem.mem.DecodeZStringAt(tmpAddr, header)
 
 				alphabet = shiftLock
@@ -149,12 +149,12 @@ func (zmem *ZMemorySequential) DecodeZString(header *ZHeader) string {
 	return ret
 }
 
-func PackedAddress(addr uint16) uint16 {
+func PackedAddress(addr uint32) uint32 {
 	// v3
 	return addr * 2
 }
 
-func IsPackedAddress(addr uint16) bool {
+func IsPackedAddress(addr uint32) bool {
 	// v3
 	return addr%2 == 0
 }
